code cleanup
git-svn-id: https://cida-svn.er.usgs.gov/repos/dev/usgs/gdp/trunk@10307 622dc1fb-2224-0410-ab72-dc85a3975e1a

+++ src/main/java/gov/usgs/gdp/servlet/WCSServlet.java
@@ -141,7 +141,7 @@ public class WCSServlet extends HttpServlet {
 			
 			if (size > MAX_COVERAGE_SIZE) {
 				float percent = (float) size / MAX_COVERAGE_SIZE * 100 - 100;
-				String percentString = String.format("%1$.1f", percent);
+				String percentString = String.format("%1$.1f", Float.valueOf(percent));
 				
 				sendFailReply(response, "Coverage exceeds size limit by " + 
 						percentString + "%");
@@ -153,11 +153,11 @@ public class WCSServlet extends HttpServlet {
 	
 	void sendFailReply(HttpServletResponse response, String message) throws IOException {
 		XmlReplyBean xmlReply = new XmlReplyBean(AckBean.ACK_FAIL, new ErrorBean(message));
-		RouterServlet.sendXml(xmlReply, new Date().getTime(), response);
+		RouterServlet.sendXml(xmlReply, Long.valueOf(new Date().getTime()), response);
 	}
 	
 	void sendOkReply(HttpServletResponse response) throws IOException {
 		XmlReplyBean xmlReply = new XmlReplyBean(AckBean.ACK_OK);
-		RouterServlet.sendXml(xmlReply, new Date().getTime(), response);
+		RouterServlet.sendXml(xmlReply,Long.valueOf(new Date().getTime()), response);
 	}
 }
